requirements.txt
# ═══════════════════════════════════════════════════════════════
#           NEXUS 10 AI AGENCY - Dependencies
# ═══════════════════════════════════════════════════════════════

# === Core AI Framework ===
crewai>=0.28.0
crewai-tools>=0.2.0

# === LLM Integration ===
langchain-openai>=0.1.0
langchain-community>=0.0.10
openai>=1.0.0

# === Web Search (Free) ===
duckduckgo-search>=4.0.0
wikipedia>=1.4.0

# === Web Framework ===
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# === Telegram Bot ===
pyTelegramBotAPI>=4.14.0

# === Data Processing ===
pandas>=2.0.0

# === PDF Generation ===
fpdf2>=2.7.0
reportlab>=4.0.0

# === Environment Management ===
python-dotenv>=1.0.0

# === HTTP Requests ===
requests>=2.31.0

# === Image Processing ===
Pillow>=10.0.0

# === Scheduler ===
APScheduler>=3.10.0

# === Database ===
# SQLite included in Python stdlib

# === Optional: Better terminal output ===
rich>=13.0.0

# === Optional: Streamlit UI ===
streamlit>=1.30.0

# === Optional: Vector Store ===
chromadb>=0.4.0
